Fix questions React component behaviour
