--- app/javascript/components/QuestionsForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./QuestionsForm";
