+++ app/javascript/models/store.ts
@@ -1,5 +1,6 @@
 import { types } from "mobx-state-tree";
 import question from "./question";
+import {observable} from "mobx";
 
 const store = types
   .model({
@@ -7,21 +8,43 @@ const store = types
     current_question: types.optional(types.string, ""),
     questionsList: types.optional(types.map(question), {})
   })
+  .views(self => ({
+
+  }))
   .actions(self => ({
     addQuestion(questionData) {
-      self.questionsList.put(question.create(questionData));
+      self.questionsList.set(questionData.id, question.create({
+        id: questionData.id,
+        content: questionData.content,
+        author_id: questionData.author_id,
+        created_at: new Date(questionData.created_at)
+      });
     },
     setQuestionsList(questionsListData) {
       self.questionsList.clear();
       questionsListData.forEach(questionData => {
-        self.questionsList.put(question.create(questionData));
+        self.questionsList.set(questionData.id, question.create({
+          id: questionData.id,
+          content: questionData.content,
+          author_id: questionData.author_id,
+          created_at: new Date(questionData.created_at)
+        });
       });
     },
     clearCurrentQuestion() {
       self.current_question = "";
     },
+    setCurrentQuestion(content) {
+      self.current_question = content.replace(/[\r\n\v]+/g, "");
+    },
     getCurrentQuestion() {
       return self.current_question.trim();
+    },
+    getQuestionById(id) {
+      return self.questionsList.get(id);
+    },
+    getQuestionsList() {
+      //return self.questionsList.
     }
     // sortQuestionsList() {
     //   // sort questions by creation date to prevent wrong sorting
